fix(api): prevent duplicate registration of dev routes; add test
